Update codeowners and cargo vet README (#2)
This PR updates the CODEOWNERS file to use the EC code owners team as
the default codeowners for the repository.
This hardens the repository against stale permission issues down the
line by tying the access to being part of the team rather than being
explicitly granted permission to a repository.

This PR also updates the README for cargo vet with more details about
audit criteria