Add common traits for real-time clock hardware (#4)
This change introduces a set of traits for exposing some of the
functionality of real-time clock hardware, specifically a 'datetime
clock' and NVRAM storage.

To support the datetime clock, the Datetime type is adapted from
embassy-imxrt with the intention of removing it from over there once
this goes in. As part of that adaptation, made the datetime type fully
const.