Make RTC traits more usable in dyn contexts (#5)
Currently, the RTC traits are not easily usable in dyn contexts. The
DatetimeClock trait has an associated constant, which is inherently not
dyn-safe, and all the traits do not have Send as a supertrait, which
means that attempting to use them in static dyn contexts requires
specifying `T + Send` everywhere.

I believe there are no reasonable implementations of these traits that
are not Send, so this change adds that as a supertrait to all HAL traits
and changes DatetimeClock to use a member function rather than an
associated constant.